+++ rambdax.js
@@ -1,4 +1,3 @@
-export * from './src/contains'
 export * from './src/allFalse'
 export * from './src/allTrue'
 export * from './src/allType'